# src/api/openapi.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package api

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// ObjectReference defines model for ObjectReference.
type ObjectReference struct {
	// Id Identifier of the created or existing object.
	Id string `json:"id"`
}

// Bucket defines model for Bucket.
type Bucket = string

// ObjectID defines model for ObjectID.
type ObjectID = string

// CreateObjectTextBody defines parameters for CreateObject.
type CreateObjectTextBody = string

// CreateObjectTextRequestBody defines body for CreateObject for text/plain ContentType.
type CreateObjectTextRequestBody = CreateObjectTextBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Delete an object
	// (DELETE /objects/{bucket}/{objectID})
	DeleteObject(w http.ResponseWriter, r *http.Request, bucket Bucket, objectID ObjectID)
	// Download an object
	// (GET /objects/{bucket}/{objectID})
	GetObject(w http.ResponseWriter, r *http.Request, bucket Bucket, objectID ObjectID)
	// Upload an object
	// (PUT /objects/{bucket}/{objectID})
	CreateObject(w http.ResponseWriter, r *http.Request, bucket Bucket, objectID ObjectID)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// DeleteObject operation middleware
func (siw *ServerInterfaceWrapper) DeleteObject(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "bucket" -------------
	var bucket Bucket

	err = runtime.BindStyledParameterWithOptions("simple", "bucket", r.PathValue("bucket"), &bucket, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bucket", Err: err})
		return
	}

	// ------------- Path parameter "objectID" -------------
	var objectID ObjectID

	err = runtime.BindStyledParameterWithOptions("simple", "objectID", r.PathValue("objectID"), &objectID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "objectID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteObject(w, r, bucket, objectID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetObject operation middleware
func (siw *ServerInterfaceWrapper) GetObject(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "bucket" -------------
	var bucket Bucket

	err = runtime.BindStyledParameterWithOptions("simple", "bucket", r.PathValue("bucket"), &bucket, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bucket", Err: err})
		return
	}

	// ------------- Path parameter "objectID" -------------
	var objectID ObjectID

	err = runtime.BindStyledParameterWithOptions("simple", "objectID", r.PathValue("objectID"), &objectID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "objectID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetObject(w, r, bucket, objectID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateObject operation middleware
func (siw *ServerInterfaceWrapper) CreateObject(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "bucket" -------------
	var bucket Bucket

	err = runtime.BindStyledParameterWithOptions("simple", "bucket", r.PathValue("bucket"), &bucket, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "bucket", Err: err})
		return
	}

	// ------------- Path parameter "objectID" -------------
	var objectID ObjectID

	err = runtime.BindStyledParameterWithOptions("simple", "objectID", r.PathValue("objectID"), &objectID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "objectID", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateObject(w, r, bucket, objectID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/objects/{bucket}/{objectID}", wrapper.DeleteObject)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/objects/{bucket}/{objectID}", wrapper.GetObject)
	m.HandleFunc(http.MethodPut+" "+options.BaseURL+"/objects/{bucket}/{objectID}", wrapper.CreateObject)

	return m
}

type DeleteObjectRequestObject struct {
	Bucket   Bucket   `json:"bucket"`
	ObjectID ObjectID `json:"objectID"`
}

type DeleteObjectResponseObject interface {
	VisitDeleteObjectResponse(w http.ResponseWriter) error
}

type DeleteObject200Response struct {
}

func (response DeleteObject200Response) VisitDeleteObjectResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type DeleteObject404Response struct {
}

func (response DeleteObject404Response) VisitDeleteObjectResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type GetObjectRequestObject struct {
	Bucket   Bucket   `json:"bucket"`
	ObjectID ObjectID `json:"objectID"`
}

type GetObjectResponseObject interface {
	VisitGetObjectResponse(w http.ResponseWriter) error
}

type GetObject200TextResponse string

func (response GetObject200TextResponse) VisitGetObjectResponse(w http.ResponseWriter) error {

	w.Header().Set("Content-Type", "text/plain")
	w.WriteHeader(200)

	_, err := w.Write([]byte(response))
	return err
}

type GetObject404Response struct {
}

func (response GetObject404Response) VisitGetObjectResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type CreateObjectRequestObject struct {
	Bucket   Bucket   `json:"bucket"`
	ObjectID ObjectID `json:"objectID"`
	Body     *CreateObjectTextRequestBody
}

type CreateObjectResponseObject interface {
	VisitCreateObjectResponse(w http.ResponseWriter) error
}

type CreateObject201JSONResponse ObjectReference

func (response CreateObject201JSONResponse) VisitCreateObjectResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type CreateObject409ResponseHeaders struct {
	Location string
}

type CreateObject409JSONResponse struct {
	Body    ObjectReference
	Headers CreateObject409ResponseHeaders
}

func (response CreateObject409JSONResponse) VisitCreateObjectResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.Header().Set("Location", fmt.Sprint(response.Headers.Location))
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Delete an object
	// (DELETE /objects/{bucket}/{objectID})
	DeleteObject(ctx context.Context, request DeleteObjectRequestObject) (DeleteObjectResponseObject, error)
	// Download an object
	// (GET /objects/{bucket}/{objectID})
	GetObject(ctx context.Context, request GetObjectRequestObject) (GetObjectResponseObject, error)
	// Upload an object
	// (PUT /objects/{bucket}/{objectID})
	CreateObject(ctx context.Context, request CreateObjectRequestObject) (CreateObjectResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// DeleteObject operation middleware
func (sh *strictHandler) DeleteObject(w http.ResponseWriter, r *http.Request, bucket Bucket, objectID ObjectID) {
	var request DeleteObjectRequestObject

	request.Bucket = bucket
	request.ObjectID = objectID

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteObject(ctx, request.(DeleteObjectRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteObject")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeleteObjectResponseObject); ok {
		if err := validResponse.VisitDeleteObjectResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetObject operation middleware
func (sh *strictHandler) GetObject(w http.ResponseWriter, r *http.Request, bucket Bucket, objectID ObjectID) {
	var request GetObjectRequestObject

	request.Bucket = bucket
	request.ObjectID = objectID

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetObject(ctx, request.(GetObjectRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetObject")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetObjectResponseObject); ok {
		if err := validResponse.VisitGetObjectResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreateObject operation middleware
func (sh *strictHandler) CreateObject(w http.ResponseWriter, r *http.Request, bucket Bucket, objectID ObjectID) {
	var request CreateObjectRequestObject

	request.Bucket = bucket
	request.ObjectID = objectID

	data, err := io.ReadAll(r.Body)
	if err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't read body: %w", err))
		return
	}
	body := CreateObjectTextRequestBody(data)
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreateObject(ctx, request.(CreateObjectRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateObject")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreateObjectResponseObject); ok {
		if err := validResponse.VisitCreateObjectResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}